requirements.txt
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.57.0
fpdf2==2.8.2
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.2.1
prompt_toolkit==3.0.51
pyparsing==3.2.3
python-dateutil==2.9.0.post0
questionary==2.1.0
scapy==2.6.1
six==1.17.0
tqdm==4.67.1
wcwidth==0.2.13
